requirements.txt
haystack-ai
unstructured[github]
farm-haystack